i2c-bfin-twi: work around illegal i2c bus lock with long transfers
For transfer counts > 255 bytes, the Blackfin TWI peripheral sets the
data transfer counter DCNT to 0xFF indicating unlimited transfers.  It
then uses a flag iface->manual_stop to manually issue the STOP condition
once the required amount of bytes are received.

It seems that if the STOP condition is issued with a full receive fifo,
the TWI logic constantly drives the SDA/SCL lines low thus locking the
bus.  So to make sure this doesn't happen, consume the fifo first.

Signed-off-by: Michael Hennerich <michael.hennerich@analog.com>
Signed-off-by: Mike Frysinger <vapier@gentoo.org>