re-order CONFIG_EXPERT options to fix menuconfig display
The kconfig language requires that dependent options all follow the
menuconfig symbol in order to be collapsed below it.  Recently some
hidden options were added below the EXPERT menuconfig, but did not
depend on EXPERT (because hidden options can't).  This broke the
display.  So re-order all these options, and while we're here stick
the PCI quirks under the EXPERT menu (since it isn't sitting with
any related options).

Before this commit, we get:
	[*] Configure standard kernel features (expert users)  --->
	[ ] Sysctl syscall support
	[*] Load all symbols for debugging/ksymoops
	...
	[ ] Embedded system

Now we get the older (and correct) behavior:
	[*] Configure standard kernel features (expert users)  --->
	[ ] Embedded system
And if you go into the expert menu you get the expert options:
	[ ] Sysctl syscall support
	[*] Load all symbols for debugging/ksymoops
	...

Acked-by: Randy Dunlap <rdunlap@infradead.org>
Signed-off-by: Mike Frysinger <vapier@gentoo.org>