UBI: document UBI_IOCVOLUP better in user header
The current ioctl define implies that this func expects to be passed a
64bit number directly rather than a pointer to a 64bit.  The code that
processes this ioctl shows that it clearly expects a pointer.

It'd be best if we could change the type to "__s64*", but that would
change the generated ioctl number thus breaking the userland ABI.  So
just add a comment for intrepid developers.

Signed-off-by: Mike Frysinger <vapier@gentoo.org>